+++ src/test/java/hexlet/code/util/ModelGenerator.java
@@ -56,7 +56,7 @@ public class ModelGenerator {
                 .ignore(Select.field(Task::getTaskStatus))
                 .ignore(Select.field(Task::getLabels))
                 .supply(Select.field(Task::getName), () -> faker.lorem().word())
-                .supply(Select.field(Task::getIndex), () -> (int) faker.number().randomNumber())
+                .supply(Select.field(Task::getIndex), () -> faker.number().randomNumber())
                 .supply(Select.field(Task::getDescription), () -> faker.lorem().sentence())
                 .toModel();
